requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob==12.9.0
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
filelock==3.6.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.4.2
parsel==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyodbc==4.0.32
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
tldextract==3.3.0
tomli==2.0.1
Twisted==22.4.0
typing-extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
zope.interface==5.4.0
